python-service/requirements.txt
flask==3.1.0
flask-cors==5.0.1
numpy==2.2.3
Pillow==11.1.0
trimesh==4.6.6
scipy==1.15.2
gunicorn==23.0.0
